Code Upload
This code is to be used in conjunction with MS Access to create a login system with permissions, segmenting users and limiting access based on role/position. It uses a small form with two text boxes and a button to send the information query out to the program. From there, you can choose to show content based on role, hiding specific tables, queries or forms from users. This draws from a table named "EmployeeDB", utilizing a name, username, password and role fields. As I created this specifically for a company, I will not be uploading the whole access project for confidentiality reasons, but I'm hoping that this VB code comes in handy to MS Access users!